requirements.txt
asgiref==3.7.2
black==23.7.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
crispy-bootstrap4==2022.1
cryptography==41.0.2
defusedxml==0.7.1
distlib==0.3.7
dj-database-url==2.0.0
Django==4.2.3
django-allauth==0.54.0
django-crispy-forms==2.0
django-environ==0.10.0
djangorestframework==3.14.0
filelock==3.12.2
gunicorn==21.2.0
idna==3.4
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.2
pipenv==2023.7.23
platformdirs==3.10.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.4
virtualenv==20.24.2
virtualenv-clone==0.5.7
whitenoise==6.5.0
